backend/requirements.txt
Flask>=2.3.0,<2.4
gunicorn
pytest==7.4.0
flake8==6.1.0
psycopg2-binary==2.9.7

Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
PyJWT==2.8.0
Werkzeug==2.3.7
requests
requests
